update with cleaned code
