+++ kaggle/RANZCR/4th_place_solution/configs/cfg_seg_40_1024d_full.py
@@ -22,7 +22,7 @@ from default_config import basic_cfg
 cfg = basic_cfg
 
 cfg.lr = 0.0006
-cfg.backbone = "tf_efficientnet_b8_ap"
+cfg.backbone = "efficientnet_b8_ap"
 cfg.img_size = (896, 896)
 cfg.grad_accumulation = 2
